Generate version header that includes build details
This now generates a 'version.hpp' file directly into the include
directory. This file is ignored by git CVS, and is used to determine
build configurations about the library; such as the size of the
floating-point values to use, and whether or not to use cached
trig.

Prior to this, a latent bug existed wherein it's possible to
build this library for different configurations than it would
be used for (e.g. built with 'double' precision, but then used
with single precision by forgetting to define 'BIT_MATH_DOUBLE_PRECISION').